+++ package.json
@@ -169,7 +169,6 @@
     "reinstall": "yarn run clean && yarn run bootstrap",
     "docs": "node --max-old-space-size=8192 ./scripts/docs/build-docs.js",
     "docs:api": "node ./scripts/docs/build-api-docs.js",
-    "release:find-plugins": "node ./scripts/docs/find-new-plugins.js",
     "docs:serve": "http-server ./build/docs/",
     "docs:verify": "node ./scripts/web-crawler/index.js --docs",
     "translations:collect": "node ./scripts/translations/collect.js",
@@ -179,10 +178,8 @@
     "predll:build": "npm run build",
     "dll:build": "node ./scripts/dll/build-dlls.js --base-dll-config ./scripts/dll/webpack.config.dll.js",
     "changelog": "node ./scripts/release/changelog.js",
-    "postchangelog": "node ./scripts/release/update-version-references.js",
-    "release:bump-version": "node ./scripts/release/bump-versions.js",
-    "postrelease:bump-version": "npm run dll:build",
-    "release:publish": "node ./scripts/release/publish.js",
+    "release:find-plugins": "node ./scripts/docs/find-new-plugins.js",
+    "release:prepare-packages": "node ./scripts/release/preparepackages.js",
     "release:clean": "node ./scripts/release/clean.js",
     "clean-up-svg-icons": "sh ./scripts/clean-up-svg-icons.sh",
     "collect-svg-icons": "node scripts/collect-svg-icons.js",
