+++ scripts/release/preparepackages.js
@@ -9,33 +9,37 @@
 
 'use strict';
 
-const fs = require( 'fs' );
-const upath = require( 'upath' );
+const { EventEmitter } = require( 'events' );
 const releaseTools = require( '@ckeditor/ckeditor5-dev-release-tools' );
 const { Listr } = require( 'listr2' );
-const updateVersionReferences = require( './update-version-references' );
-const buildTsAndDllForCkeditor5Root = require( './buildtsanddllforckeditor5root' );
+const updateVersionReferences = require( './utils/updateversionreferences' );
+const buildTsAndDllForCkeditor5Root = require( './utils/buildtsanddllforckeditor5root' );
+const getCKEditor5PackageJson = require( './utils/getckeditor5packagejson' );
+const parseArguments = require( './utils/parsearguments' );
+const isCKEditor5Package = require( './utils/isckeditor5package' );
+const compileTypeScriptCallback = require( './utils/compiletypescriptcallback' );
+const updatePackageEntryPoint = require( './utils/updatepackageentrypoint' );
+const prepareDllBuildsCallback = require( './utils/preparedllbuildscallback' );
 
-// TODO: Integrate with `minimist` to read CLI arguments.
+const cliArguments = parseArguments( process.argv.slice( 2 ) );
+
+// `executeInParallel()` is executed thrice.
+EventEmitter.defaultMaxListeners = ( cliArguments.concurrency * 3 + 1 );
 
 const abortController = new AbortController();
 const PACKAGES_DIRECTORY = 'packages';
 const RELEASE_DIRECTORY = 'release';
 const latestVersion = releaseTools.getLastFromChangelog();
 
+const taskOptions = {
+	rendererOptions: {
+		collapseSubtasks: false
+	}
+};
+
 process.on( 'SIGINT', () => {
 	abortController.abort( 'SIGINT' );
 } );
-
-process.on( 'unhandledRejection', error => {
-	console.error( error );
-
-	process.exit( 1 );
-} );
-
-abortController.signal.addEventListener( 'abort', () => {
-	process.exit( 1 );
-}, { once: true } );
 
 const tasks = new Listr( [
 	{
@@ -48,7 +52,7 @@ const tasks = new Listr( [
 						return releaseTools.updateDependencies( {
 							version: '^' + latestVersion,
 							packagesDirectory: PACKAGES_DIRECTORY,
-							shouldUpdateVersionCallback: require( './isckeditor5package' )
+							shouldUpdateVersionCallback: isCKEditor5Package
 						} );
 					}
 				},
@@ -70,7 +74,7 @@ const tasks = new Listr( [
 						} );
 					}
 				}
-			] );
+			], taskOptions );
 		}
 	},
 	{
@@ -84,23 +88,32 @@ const tasks = new Listr( [
 					}
 				},
 				{
+					title: 'Prepare "ckeditor5-build-*" builds.',
+					task: () => {
+						// TODO: Waits for #14177.
+						return Promise.resolve();
+					}
+				},
+				{
 					title: 'Compile TypeScript in `ckeditor5-*` packages.',
 					task: ( ctx, task ) => {
 						return releaseTools.executeInParallel( {
 							packagesDirectory: PACKAGES_DIRECTORY,
 							signal: abortController.signal,
 							listrTask: task,
-							taskToExecute: require( './compiletypescriptcallback' )
+							taskToExecute: compileTypeScriptCallback,
+							concurrency: cliArguments.concurrency
 						} );
 					}
 				},
 				{
-					title: '',
+					title: 'Copying CKEditor 5 packages.',
 					task: () => {
 						return releaseTools.prepareRepository( {
 							outputDirectory: RELEASE_DIRECTORY,
 							packagesDirectory: PACKAGES_DIRECTORY,
-							rootPackageJson: getCKEditor5PackageJson()
+							rootPackageJson: getCKEditor5PackageJson(),
+							packagesToCopy: cliArguments.packages
 						} );
 					}
 				},
@@ -111,7 +124,8 @@ const tasks = new Listr( [
 							packagesDirectory: RELEASE_DIRECTORY,
 							signal: abortController.signal,
 							listrTask: task,
-							taskToExecute: require( './updatepackageentrypoint' )
+							taskToExecute: updatePackageEntryPoint,
+							concurrency: cliArguments.concurrency
 						} );
 					}
 				},
@@ -122,11 +136,12 @@ const tasks = new Listr( [
 							packagesDirectory: RELEASE_DIRECTORY,
 							signal: abortController.signal,
 							listrTask: task,
-							taskToExecute: require( './preparedllbuildscallback' )
+							taskToExecute: prepareDllBuildsCallback,
+							concurrency: cliArguments.concurrency
 						} );
 					}
 				}
-			] );
+			], taskOptions );
 		}
 	},
 	{
@@ -150,74 +165,9 @@ const tasks = new Listr( [
 			} );
 		}
 	}
-] );
+], taskOptions );
 
 tasks.run()
 	.catch( err => {
 		console.error( err );
 	} );
-
-/**
- * @returns {Object}
- */
-function getCKEditor5PackageJson() {
-	const pkgJson = require( '../../package.json' );
-
-	return {
-		name: pkgJson.name,
-		version: pkgJson.version,
-		keywords: pkgJson.keywords,
-		description: 'A set of ready-to-use rich text editors created with a powerful framework.' +
-			' Made with real-time collaborative editing in mind.',
-		dependencies: getCKEditor5Dependencies( pkgJson.dependencies ),
-		engines: pkgJson.engines,
-		author: pkgJson.author,
-		license: pkgJson.license,
-		homepage: pkgJson.homepage,
-		bugs: pkgJson.bugs,
-		repository: pkgJson.repository,
-		files: [
-			// Do not add the entire `build/` directory as it contains files produced by internal scripts:
-			// automated/manual tests, translations, documentation, content styles.
-			// If you need to release anything from the directory, insert a relative path to the file/directory.
-			'src/*.js',
-			'src/*.d.ts',
-			'build/ckeditor5-dll.js',
-			'build/ckeditor5-dll.manifest.json',
-			'build/translations/*.js',
-			// npm default files.
-			'CHANGELOG.md',
-			'LICENSE.md',
-			'README.md'
-		]
-	};
-}
-
-/**
- * Returns an array that contains name of packages that the `ckeditor5` package should define as its dependencies.
- *
- * @param {Object} dependencies Dependencies to filter out.
- * @returns {Array.<String>}
- */
-function getCKEditor5Dependencies( dependencies ) {
-	// Short name of packages specified as DLL.
-	const dllPackages = fs.readdirSync( upath.join( __dirname, '..', '..', 'src' ) )
-		.map( directory => directory.replace( /\.[tj]s$/, '' ) );
-
-	// Name of packages that are listed in `src/` as DLL packages.
-	const ckeditor5Dependencies = Object.keys( dependencies )
-		.filter( packageName => {
-			const shortPackageName = packageName.replace( /@ckeditor\/ckeditor5?-/, '' );
-
-			return dllPackages.includes( shortPackageName );
-		} );
-
-	// The proper object for inserting into the `package.json` file.
-	const dependencyObject = {};
-
-	for ( const item of ckeditor5Dependencies ) {
-		dependencyObject[ item ] = dependencies[ item ];
-	}
-
-	return dependencyObject;
-}
