+++ scripts/release/utils/updatepackageentrypoint.js
