+++ scripts/release/utils/updateversionreferences.js
